[Refactor/#60] 등록 관련 data 모델 수정
