+++ app/src/main/java/com/poti/android/data/remote/datasource/ArtistRemoteDataSource.kt
@@ -1,7 +1,6 @@
 package com.poti.android.data.remote.datasource
 
-import com.poti.android.data.remote.dto.base.ApiResponse
-import com.poti.android.data.remote.dto.response.MemberListResponseDto
+import com.poti.android.data.remote.dto.response.artist.MemberListResponseDto
 import com.poti.android.core.network.model.BaseResponse
 import com.poti.android.data.remote.dto.response.artist.ArtistListResponseDto
 import com.poti.android.data.remote.service.ArtistService
@@ -13,6 +12,6 @@ class ArtistRemoteDataSource @Inject constructor(
     suspend fun getArtists(): BaseResponse<ArtistListResponseDto> =
         artistService.getArtists()
 
-    suspend fun getMemberList(artistId: Long): ApiResponse<MemberListResponseDto> =
+    suspend fun getMemberList(artistId: Long): BaseResponse<MemberListResponseDto> =
         artistService.getMemberList(artistId)
 }
