--- app/src/main/java/com/poti/android/data/remote/dto/base/ApiResponse.kt
@@ -1,14 +1,0 @@
-package com.poti.android.data.remote.dto.base
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-@Serializable
-data class ApiResponse<T>(
-    @SerialName("code")
-    val code: Int,
-    @SerialName("message")
-    val message: String,
-    @SerialName("data")
-    val data: T
-)
