+++ app/src/main/java/com/poti/android/data/remote/dto/response/artist/ArtistSearchListResponse.kt
@@ -1,0 +1,18 @@
+package com.poti.android.data.remote.dto.response.artist
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class ArtistSearchListResponseDto(
+    @SerialName("artists")
+    val artists: List<ArtistSearchResponseDto>,
+)
+
+@Serializable
+data class ArtistSearchResponseDto(
+    @SerialName("artists")
+    val artists: Long,
+    @SerialName("name")
+    val name: String,
+)
