+++ app/src/main/java/com/poti/android/data/remote/service/ArtistService.kt
@@ -1,7 +1,6 @@
 package com.poti.android.data.remote.service
 
-import com.poti.android.data.remote.dto.base.ApiResponse
-import com.poti.android.data.remote.dto.response.MemberListResponseDto
+import com.poti.android.data.remote.dto.response.artist.MemberListResponseDto
 import com.poti.android.core.network.model.BaseResponse
 import com.poti.android.data.remote.dto.response.artist.ArtistListResponseDto
 import retrofit2.http.GET
@@ -11,7 +10,7 @@ interface ArtistService {
     @GET("api/v1/artists/{artistId}/members")
     suspend fun getMemberList(
         @Path("artistId") artistId: Long,
-    ): ApiResponse<MemberListResponseDto>
+    ): BaseResponse<MemberListResponseDto>
 
     @GET("/api/v1/artists")
     suspend fun getArtists(): BaseResponse<ArtistListResponseDto>
