+++ app/src/main/java/com/poti/android/data/remote/service/ImageService.kt
@@ -1,7 +1,7 @@
 package com.poti.android.data.remote.service
 
-import com.poti.android.data.remote.dto.base.ApiResponse
-import com.poti.android.data.remote.dto.response.PresignedUrlListResponseDto
+import com.poti.android.core.network.model.BaseResponse
+import com.poti.android.data.remote.dto.response.image.PresignedUrlListResponseDto
 import retrofit2.http.GET
 import retrofit2.http.Query
 
@@ -11,5 +11,5 @@ interface ImageService {
         @Query("type") type: String,
         @Query("count") count: Int,
         @Query("extension") extension: String,
-    ): ApiResponse<PresignedUrlListResponseDto>
+    ): BaseResponse<PresignedUrlListResponseDto>
 }
